% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4_integration_funs.R
\name{get_Psi}
\alias{get_Psi}
\title{Get the covariance matrix of random effects}
\usage{
get_Psi(lmerfit, psi = NULL)
}
\arguments{
\item{lmerfit}{An lmerMod object from fitting a linear mixed model using lme4::lmer}

\item{psi}{Optional vector with covariance parameters (see details)}
}
\value{
A usually sparse covariance matrix of random effects of type dsCMatrix
}
\description{
Returns the covariance matrix of the random effects in a linear mixed model,
either as estimated by lme4::lmer or evaluated at a particular parameter value
supplied as an argument.
}
\details{
If psi is not supplied, the estimated covariance matrix is returned. If psi is
supplied, the covariance matrix is calculated using those parameter values instead.
psi should be NULL (default) or a numeric vector of length getME(lmerfit, "m").
In the latter case the elements should be ordered as those in the vcov column of
as.data.frame(VarCorr(lmerfit), order = "lower.tri"). In particular, the last
element is the error variance.
}
