% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_PsiRcpp}
\alias{get_PsiRcpp}
\title{get_PsiRcpp}
\usage{
get_PsiRcpp(psi, H)
}
\arguments{
\item{H}{Matrix of derivatives of Psi with respect to elements of psi.
Assumes H = \link{H_1, ... , H_r}, where H_j is q by q.}

\item{psi1}{Covariance parameter vector}
}
\value{
The covariance matrix Psi
}
\description{
Constructs the covariance matrix from the covariance parameters
This function is implemented in C++ and is faster than the equivalent function \code{get_Psi}
}
