% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_funs.R
\name{loglik_psi}
\alias{loglik_psi}
\title{loglik_psi}
\usage{
loglik_psi(
  Z,
  ZtZXe,
  e,
  H,
  Psi_r,
  psi_r,
  loglik = TRUE,
  score = TRUE,
  finf = TRUE,
  expected = TRUE
)
}
\arguments{
\item{Z}{A matrix of fixed effects.}

\item{ZtZXe}{Pre-computed matrix with of \eqn{Z'[Z, X, e]}}

\item{e}{The residual vector.}

\item{H}{Matrix of derivatives of Psi with respect to elements of psi.
Assumes \eqn{H = [H_1, ... , H_{r - 1}]}, where \eqn{H_j} is q by q.}

\item{Psi_r}{Covariance matrix of random effects (Psi) divided by error
variance psi_r, with dimensions q by q.}

\item{psi_r}{The error or variance.}

\item{loglik}{If \code{TRUE} (default), the log-likelihood will be calculated.}

\item{score}{If \code{TRUE} (default), the score vector will be calculated.}

\item{finf}{If \code{TRUE} (default), the Fisher information matrix will be calculated.}

\item{expected}{If \code{TRUE} (detault), return expected information;
otherwise observed.}
}
\value{
A list with components:
\item{ll}{The log-likelihood.}
\item{score}{The score vector.}
\item{finf}{The Fisher information matrix.}
}
\description{
Calculates the log-likelihood, score vector and Fisher information matrix
for the covariance parameter vector \code{psi} in a linear mixed effects model.
}
