% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_res}
\alias{loglik_res}
\title{Restricted log-likelihood using RcppEigen}
\usage{
loglik_res(
  Psi_r,
  psi_r,
  H,
  Y,
  X,
  Z,
  XtX,
  XtZ,
  ZtZ,
  XtY,
  ZtY,
  get_val = TRUE,
  get_score = TRUE,
  get_inf = TRUE
)
}
\arguments{
\item{Psi_r}{The \eqn{q\times q} covariance matrix of random effects (\eqn{\Psi}) divided by error
variance, \eqn{\Psi_r = \Psi / \psi_r} of class \code{dgCMatrix}.}

\item{psi_r}{The error variance \eqn{\psi_r > 0}.}

\item{H}{Sparse \eqn{q \times q(r - 1)} matrix of horizontally concatenated
derivatives of \eqn{\Psi} (see details) of class \code{dgCMatrix}.}

\item{Y}{Vector of length \eqn{n} of responses, of class \code{numeric}.}

\item{X}{Matrix of size \eqn{n \times p} of predictors, of class \code{matrix}.}

\item{Z}{Sparse \eqn{n \times q} random effect design matrix of class \code{dgCMatrix}.}

\item{XtX}{Precomputed matrix \code{crossprod(X)} of class \code{matrix}.}

\item{XtZ}{Precomputed matrix \code{crossprod(X, Z)} of class \code{matrix}.}

\item{ZtZ}{Precomputed matrix \code{crossprod(Z)} of class \code{dgCMatrix}.}

\item{XtY}{Precomputed vector \code{crossprod(X, Y)} of class \code{numeric}.}

\item{ZtY}{Precomputed vector \code{crossprod(Z, Y)} of class \code{numeric}.}

\item{get_val}{If \code{TRUE}, the value of the loglikelihood is computed.}

\item{get_score}{If \code{TRUE} the score vector is calculated.}

\item{get_inf}{If \code{TRUE}, an information matrix is calculated.}
}
\value{
A list with components:
\item{value}{The value of the restricted log-likelihood}
\item{score}{The restricted score, or gradient of the restricted log-likelihood, for \eqn{\psi}}
\item{inf_mat}{The restricted information matrix for \eqn{\psi}}
\item{beta}{Partial maximizer of the regular likelihood in \eqn{\beta},
\eqn{\tilde{\beta} = (X' \Sigma^{-1} X)^{-1} X' \Sigma^{-1}Y},
where \eqn{\Sigma = Z\Psi Z' + \psi_r I_n}}
\item{I_b_inv_chol}{Cholesky root of the expected inverse information matrix
for \eqn{\beta}, \eqn{I(\beta; \psi)^{-1} = (X' \Sigma^{-1} X)^{-1}}}
}
\description{
Computes the restricted log-likelihood, score vector, and information matrix
for the covariance parameter vector in a linear mixed effects model.
}
\details{
The model is \deqn{Y = X\beta + Z U + E,} where \eqn{U \sim N_q(0, \Psi)}
and \eqn{E \sim N_n(0, \psi_r I_n)}. The first \eqn{r - 1} elements of \eqn{\psi}
parameterize \eqn{\Psi}, while the \eqn{r}th and last element is the error
variance. It is assumed that \eqn{H_j = \partial \Psi / \partial \psi_j} is
a (usually sparse) matrix of zeros and ones, \eqn{j \in \{1, \dots, r - 1\}},
and that \eqn{\Psi = \sum_{j = 1}^{r - 1}\psi_j H_j}. Thus, \eqn{\psi_1, \dots, \psi_{r - 1}}
are variances and covariances of random effects.
The argument matrix \code{H} is \eqn{H = [H_1, \dots, H_{r - 1}]}.

The restricted likelihood integrates out the fixed effects \eqn{\beta}.
}
