% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Psi_from_H_cpp}
\alias{Psi_from_H_cpp}
\title{Psi_from_H_cpp}
\usage{
Psi_from_H_cpp(psi_mr, H)
}
\arguments{
\item{psi_mr}{A vector of covariance parameter (see ?loglikelihood)}

\item{H}{Sparse matrix of derivatives of Psi with respect to elements of psi,
\eqn{H = [H_1, \dots , H_{r - 1}]}, where \eqn{H_j = \partial \Psi / \partial \psi_j}.}
}
\value{
The covariance matrix \eqn{\Psi}
}
\description{
Constructs the covariance matrix of the random effects
}
