% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_funs.R
\name{res_ll}
\alias{res_ll}
\title{Restricted likelihood in pure R}
\usage{
res_ll(
  XtX,
  XtY,
  XtZ,
  ZtZ,
  YtZ,
  Y,
  X,
  Z,
  H,
  Psi_r,
  psi_r,
  get_val = TRUE,
  get_score = FALSE,
  get_inf = FALSE
)
}
\arguments{
\item{XtX}{An n x p matrix of the crossproduct of the design matrix of fixed
effects (\eqn{X}) with itself.}

\item{XtY}{An \eqn{p \times 1} vector of the crossproduct of the design matrix of fixed
effects (\eqn{X}) with the response vector (\eqn{Y}).}

\item{XtZ}{An \eqn{p \teims q} matrix of the crossproduct of the design matrix of fixed
effects (\eqn{X}) with the design matrix of random effects (\eqn{Z}).}

\item{ZtZ}{A q x q matrix of the crossproduct of the design matrix of random
effects (\eqn{Z}) with itself.}

\item{YtZ}{A 1 x q matrix of the crossproduct of the response vector (\eqn{Y}) with
the design matrix of random effects (\eqn{Z}).}

\item{Y}{An n x 1 vector of the response variable.}

\item{X}{An n x p matrix of the design matrix of fixed effects.}

\item{Z}{An n x q matrix of the design matrix of random effects.}

\item{H}{A q x rq matrix, where \eqn{H = [H_1, \dots , H_r]}, with \eqn{H_j = \partial \Psi / \partial \psi_j}}

\item{Psi_r}{The covariance matrix of the random effects (\eqn{Psi}) divided by the
error variance (\eqn{\psi_r})}

\item{psi_r}{A scalar value of the error variance.}

\item{get_val}{If \code{TRUE} (default), the log-likelihood will be computed.}

\item{get_score}{If \code{TRUE} (default), the score vector will be computed.}

\item{get_inf}{If \code{TRUE} (default), the Fisher information matrix will be
computed.}
}
\value{
A list with elements:
\item{value}{The restricted log-likelihood at supplied parameters.}
\item{score}{The restricted score at the supplied parameters.}
\item{inf_mat}{The expected restricted information matrix at supplied parameters}
}
\description{
Computes the restricted (residual) likelihood, score, and information matrix
for the variance parameter \code{psi} a linear mixed effects model
}
