% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{res_llRcpp}
\alias{res_llRcpp}
\title{Compute Restricted Likelihood, Score and Information}
\usage{
res_llRcpp(X, Y, Z, H, Psi0, psi0, lik = TRUE, score = FALSE, finf = FALSE)
}
\arguments{
\item{X}{An n x p matrix of the design matrix of fixed effects.}

\item{Y}{An n x 1 vector of the response variable.}

\item{Z}{An n x q matrix of the design matrix of random effects.}

\item{H}{A q x rq matrix, where H = \link{H_1, ..., H_r}, with H_j being the
derivative of Psi with respect to psi_j}

\item{Psi0}{The covariance matrix of the random effects (Psi) divided by the
error variance (psi0)}

\item{psi0}{A scalar value of the error variance.}

\item{lik}{If \code{TRUE} (default), the log-likelihood will be computed.}

\item{score}{If \code{TRUE} (default), the score vector will be computed.}

\item{finf}{If \code{TRUE} (default), the Fisher information matrix will be
computed.}
}
\value{
A list with elements:
\describe{
\item{ll}{A scalar value of the restricted log-likelihood.}
\item{score}{A (r + 1) x 1 vector of the restricted score of the variance parameters.}
\item{finf}{A (r + 1) x (r + 1) matrix of the restricted  information of the
variance parameters.}
\item{beta}{}
\item{I_b_inv_chol}{}
}
}
\description{
Computes the restricted (residual) likelihood, score, and information matrix
for the variance parameter \code{psi} a linear mixed effects model
}
