% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loglik_psi}
\alias{loglik_psi}
\title{loglik_psi}
\usage{
loglik_psi(
  Z,
  ZtZXe,
  e,
  H,
  Psi0,
  psi0,
  loglik = TRUE,
  score = TRUE,
  finf = TRUE
)
}
\arguments{
\item{Z}{A matrix of fixed effects.}

\item{ZtZXe}{Pre-computed matrix with of Z'\link{Z, X, e}}

\item{e}{The residual vector.}

\item{H}{Matrix of derivatives of Psi with respect to elements of psi.
Assumes H = \link{H_1, ... , H_r}, where H_j is q by q.}

\item{Psi0}{Coavariance matrix of random effects (Psi) divided by error
variance psi0, with dimensions q by q.}

\item{psi0}{The error variance.}

\item{loglik}{If \code{TRUE} (default), the log-likelihood will be calculated.}

\item{score}{If \code{TRUE} (default), the score vector will be calculated.}

\item{finf}{If \code{TRUE} (default), the Fisher information matrix will be calculated.}
}
\value{
A list with components:
\item{ll}{The log-likelihood.}
\item{score}{The score vector.}
\item{finf}{The Fisher information matrix.}
}
\description{
Calculates the log-likelihood, score vector and Fisher information matrix
for the variance parameter vector \code{psi} in a linear mixed effects model.
}
