% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_funs.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-likelihood}
\usage{
loglikelihood(
  psi,
  b = NULL,
  Y,
  X,
  Z,
  Hlist,
  REML = TRUE,
  get_val = TRUE,
  get_score = TRUE,
  get_inf = TRUE,
  expected = TRUE
)
}
\arguments{
\item{psi}{Vector of length \eqn{r} of covariance parameters (see details)}

\item{b}{Vector of length\eqn{p} of fixed effects parameters, or regression coefficients}

\item{Y}{Vector of length \eqn{n} of responses}

\item{X}{Dense matrix of size \eqn{n\times p} of regressors}

\item{Z}{Sparse design matrix for the random effects of size \eqn{n\times q}}

\item{Hlist}{A list of derivatives of the random effects' covariance matrix \eqn{\Psi} (see details)}

\item{REML}{If \code{TRUE}, use the restricted likelihood; otherwise the regular likelihood is used}

\item{get_val}{If \code{TRUE}, the value of the log-likelihood is computed}

\item{get_score}{If \code{TRUE} the score vector, or gradient of log-likelihood, is calculated.}

\item{get_inf}{If \code{TRUE}, an information matrix is calculated.}

\item{expected}{If \code{TRUE}, the expected information is calculated; otherwise
the observed, or negative Hessian of the log-likelihood.}
}
\value{
A list with components:
\item{value}{The value of the log-likelihood}
\item{score}{The score, or gradient of the log-likelihood}
\item{inf_mat}{The information matrix}
}
\description{
Computes log-likelihood, score vector, and information matrix
for the covariance parameter vector in a linear mixed effects model.
}
\details{
The model is \deqn{Y = X\beta + Z U + E,} where \eqn{U \sim N_q(0, \Psi)}
and \eqn{E \sim N_n(0, \psi_r I_n)}. The first \eqn{r - 1} elements of \eqn{\psi}
parameterize \eqn{\Psi}, while the \eqn{r}th and last element is the error
variance. It is assumed that \eqn{H_j = \partial \Psi / \partial \psi_j} is
a (usually sparse) matrix of zeros and ones, \eqn{j \in \{1, \dots, r - 1\}},
and that \eqn{\Psi = \sum_{j = 1}^{r - 1}\psi_j H_j}. Thus, \eqn{\psi_1, \dots, \psi_{r - 1}}
are variances and covariances of random effects.

The argument \code{Hlist} is a list whose \eqn{j}th element is \eqn{H_j}.

The score and information for \eqn{\beta} are only computed if REML is FALSE,
otherwise the score and information for \eqn{\psi} only is returned.
}
