% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Psi_from_H_cpp}
\alias{Psi_from_H_cpp}
\title{Psi_from_H_cpp}
\usage{
Psi_from_H_cpp(psi_mr, H)
}
\arguments{
\item{psi_mr}{A vector of}

\item{H}{Matrix of derivatives of Psi with respect to elements of psi.
Assumes H = \link{H_1, ... , H_r}, where H_j is q by q.}
}
\value{
The covariance matrix Psi
}
\description{
Constructs the covariance matrix of the random effects
}
