% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{res_ll}
\alias{res_ll}
\title{Compute Restricted Likelihood, Score and Information}
\usage{
res_ll(
  XtX,
  XtY,
  XtZ,
  ZtZ,
  YtZ,
  Y,
  X,
  Z,
  H,
  Psi0,
  psi0,
  lik = TRUE,
  score = FALSE,
  finf = FALSE
)
}
\arguments{
\item{XtX}{An n x p matrix of the crossproduct of the design matrix of fixed
effects (X) with itself.}

\item{XtY}{An n x 1 vector of the crossproduct of the design matrix of fixed
effects (X) with the response vector (Y).}

\item{XtZ}{An n x q matrix of the crossproduct of the design matrix of fixed
effects (X) with the design matrix of random effects (Z).}

\item{ZtZ}{A q x q matrix of the crossproduct of the design matrix of random
effects (Z) with itself.}

\item{YtZ}{A 1 x q matrix of the crossproduct of the response vector (Y) with
the design matrix of random effects (Z).}

\item{Y}{An n x 1 vector of the response variable.}

\item{X}{An n x p matrix of the design matrix of fixed effects.}

\item{Z}{An n x q matrix of the design matrix of random effects.}

\item{H}{A q x rq matrix, where H = \link{H_1, ..., H_r}, with H_j being the
derivative of Psi with respect to psi_j}

\item{Psi0}{The covariance matrix of the random effects (Psi) divided by the
error variance (psi0)}

\item{psi0}{A scalar value of the error variance.}

\item{lik}{If \code{TRUE} (default), the log-likelihood will be computed.}

\item{score}{If \code{TRUE} (default), the score vector will be computed.}

\item{finf}{If \code{TRUE} (default), the Fisher information matrix will be
computed.}
}
\value{
A list with elements:
\describe{
\item{res_ll}{A scalar value of the restricted log-likelihood.}
\item{s_psi}{A (r + 1) x 1 vector of the restricted score of the variance parameters.}
\item{I_psi}{A (r + 1) x (r + 1) matrix of the restricted  information of the
variance parameters.}
}
}
\description{
Computes the restricted (residual) likelihood, score, and information matrix
for a linear mixed effects model
}
